requirements.txt
Flask >= 2.0.0
Flask_RESTful >= 0.3.9
RouterOS_api >= 0.18.0
python_dotenv >= 0.16.0
phonenumbers >= 8.13.47
flasgger >= 0.9.7
gunicorn >= 21.2.0
